Adding an updated versions of the redact.sh scrip
redact_standard_output.sh works with the current
host dns and mysql organization, and it is supposed
to be executed every time a wiki is created.
Unlike the original redact.sh, it outputs to the
stdout to be examined before being executed with
"| mysql".

redact_centralauth.txt are the 1 time triggers to be
created on centralauth.

Bug: T104900
Change-Id: I0a301cfad49d201b0e1e343e9efcd6946e483415